requirements.txt
amqp==5.1.1
anyio==3.6.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cssselect==1.1.0
Deprecated==1.2.13
Django==4.0.6
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-filter==22.1
django-redis==5.2.0
django-rest-swagger==2.2.0
django-timezone-field==5.0
djangorestframework==3.13.1
docopt==0.6.2
fake-useragent==0.1.11
flower==1.1.0
greenlet==1.1.3
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
humanize==4.2.3
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
numpy==1.23.3
openapi-codec==1.3.2
outcome==1.2.0
packaging==21.3
pandas==1.5.0
parsel==1.6.0
pipreqs==0.4.11
prometheus-client==0.14.1
prompt-toolkit==3.0.30
pycodestyle==2.8.0
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
requests==2.28.1
rfc3986==1.5.0
selenium==4.3.0
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.41
sqlparse==0.4.2
toml==0.10.2
tornado==6.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.10
uvicorn==0.20.0
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
wrapt==1.14.1
wsproto==1.1.0
yarg==0.1.9
amqp==5.1.1
anyio==3.6.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cssselect==1.1.0
Deprecated==1.2.13
Django==4.0.6
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-filter==22.1
django-redis==5.2.0
django-rest-swagger==2.2.0
django-timezone-field==5.0
djangorestframework==3.13.1
docopt==0.6.2
fake-useragent==0.1.11
flower==1.1.0
greenlet==1.1.3
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
humanize==4.2.3
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
numpy==1.23.3
openapi-codec==1.3.2
outcome==1.2.0
packaging==21.3
pandas==1.5.0
parsel==1.6.0
pipreqs==0.4.11
prometheus-client==0.14.1
prompt-toolkit==3.0.30
pycodestyle==2.8.0
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
rfc3986==1.5.0
selenium==4.3.0
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.41
sqlparse==0.4.2
toml==0.10.2
tornado==6.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.10
user-agents==2.2.0
uvicorn==0.20.0
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
wrapt==1.14.1
wsproto==1.1.0
yarg==0.1.9
